fix(feeds): validate item_count and relocate sanitiser tests (#310)
